In class exercise and almost done with assign 2

+++ assign02/src/main/java/edu/sullivb/assign02/SpaceVessel.java
@@ -19,9 +19,10 @@ public class SpaceVessel {
     public void setLength(int feet) {
         length = feet;
     }
-    public void setWeight(double pounds) {
+    public void setWeight(int pounds) {
         weight = pounds;
     }
+
     /*Getters*/
     public String getName() {
         return vesselName;
